requirements.txt
asgiref==3.6.0
branca==0.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-countries==7.5.1
django-phonenumber-field==7.1.0
folium==0.14.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
phonenumbers==8.13.11
python-dotenv==1.0.0
requests==2.30.0
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.2
whitenoise==6.4.0
